+++ chat-widget/src/common/telemetry/loggers/ariaTelemetryLogger.ts
@@ -1,14 +1,15 @@
 import { LogLevel, TelemetryInput } from "../TelemetryConstants";
-import { isNullOrEmptyString, isNullOrUndefined } from "../../utils";
+import { getDomain, isNullOrEmptyString, isNullOrUndefined } from "../../utils";
 
 import AWTEventProperties from "@microsoft/omnichannel-chat-sdk/lib/external/aria/webjs/AWTEventProperties";
 import { AWTLogConfiguration } from "@microsoft/omnichannel-chat-sdk/lib/external/aria/webjs/DataModels";
 import AWTLogManager from "@microsoft/omnichannel-chat-sdk/lib/external/aria/webjs/AWTLogManager";
 import AWTLogger from "@microsoft/omnichannel-chat-sdk/lib/external/aria/webjs/AWTLogger";
 import { AWTPiiKind } from "@microsoft/omnichannel-chat-sdk/lib/external/aria/common/Enums";
-import { Constants } from "../../Constants";
+import { Constants, AriaTelemetryConstants, EnvironmentVersion } from "../../Constants";
 import { IChatSDKLogger } from "../interfaces/IChatSDKLogger";
 import { TelemetryHelper } from "../TelemetryHelper";
+import { TelemetryManager } from "../TelemetryManager";
 
 export const ariaTelemetryLogger = (ariaTelemetryKey: string,
     disabledCookieUsage: boolean,
@@ -24,7 +25,20 @@ export const ariaTelemetryLogger = (ariaTelemetryKey: string,
 
             if (!isNullOrEmptyString(collectiorUriForTelemetry)) {
                 configuration.collectorUri = collectiorUriForTelemetry;
+            } else {
+                if (TelemetryManager.InternalTelemetryData.environmentVersion == EnvironmentVersion.prod) {
+                    const orgUrl = TelemetryManager.InternalTelemetryData?.orgUrl;
+                    if (!isNullOrUndefined(orgUrl)) {
+                        // If the given org is a Production EU org, modify the Aria collector uri
+                        const region = getDomain(orgUrl);
+
+                        if (region === AriaTelemetryConstants.EU) {
+                            configuration.collectorUri = AriaTelemetryConstants.EUROPE_ENDPOINT;
+                        }
+                    }
+                }
             }
+
             try {
                 _logger = AWTLogManager.initialize(ariaTelemetryKey, configuration);
                 if (_logger === undefined) {
@@ -39,18 +53,23 @@ export const ariaTelemetryLogger = (ariaTelemetryKey: string,
 
     const ariaLogger: IChatSDKLogger = {
         log: (logLevel: LogLevel, telemetryInput: TelemetryInput): void => {
-            let property;
-            const eventProperties = new AWTEventProperties();
-            const event = TelemetryHelper.buildTelemetryEvent(logLevel, telemetryInput);
-            eventProperties.setName(telemetryInput.scenarioType);
-            for (const key of Object.keys(event)) {
-                property = typeof (event[key]) === "object" ? JSON.stringify(event[key]) : event[key];
-                eventProperties.setProperty(key, property);
+            try {
+                let property;
+                const eventProperties = new AWTEventProperties();
+                const event = TelemetryHelper.buildTelemetryEvent(logLevel, telemetryInput);
+                eventProperties.setName(telemetryInput.scenarioType);
+                for (const key of Object.keys(event)) {
+                    property = typeof (event[key]) === "object" ? JSON.stringify(event[key]) : event[key];
+                    eventProperties.setProperty(key, property);
+                }
+                eventProperties.setPropertyWithPii(ariaTelemetryApplicationName,
+                    Constants.LiveChatWidget,
+                    AWTPiiKind.GenericData);
+                logger() ? logger().logEvent(eventProperties) : console.log("Unable to initialize aria logger");
             }
-            eventProperties.setPropertyWithPii(ariaTelemetryApplicationName,
-                Constants.LiveChatWidget,
-                AWTPiiKind.GenericData);
-            logger() ? logger().logEvent(eventProperties) : console.log("Unable to initialize aria logger");
+            catch (error) {
+                console.error("Error in logging telemetry to Aria logger:" + error);
+            }
         },
         dispose: () => {
             AWTLogManager.flush(function () { console.log("Aria logger disposed"); });