+++ CHANGE_LOG.md
@@ -16,4 +16,5 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Added custom componentOverride for preChat
 - Added Telemetry events for PostChat Loading Pane
 - Added web chat telemetry logger
+- Updated ARIA Collector URLs for EU vs Non-EU telemetry data to comply with EUDB requirements
 - Added custom context support